+++ jib-core/src/main/java/com/google/cloud/tools/jib/image/Image.java
@@ -17,6 +17,7 @@
 package com.google.cloud.tools.jib.image;
 
 import com.google.cloud.tools.jib.configuration.Port;
+import com.google.cloud.tools.jib.image.json.HistoryEntry;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import java.time.Instant;
@@ -31,13 +32,15 @@ public class Image<T extends Layer> {
   public static class Builder<T extends Layer> {
 
     private final ImageLayers.Builder<T> imageLayersBuilder = ImageLayers.builder();
-    private ImmutableMap.Builder<String, String> environmentBuilder = ImmutableMap.builder();
-    private ImmutableMap.Builder<String, String> labelsBuilder = ImmutableMap.builder();
+    private final ImmutableList.Builder<HistoryEntry> historyBuilder = ImmutableList.builder();
+    private final ImmutableMap.Builder<String, String> environmentBuilder = ImmutableMap.builder();
+    private final ImmutableMap.Builder<String, String> labelsBuilder = ImmutableMap.builder();
 
     @Nullable private Instant created;
     @Nullable private ImmutableList<String> entrypoint;
     @Nullable private ImmutableList<String> javaArguments;
     @Nullable private ImmutableList<Port> exposedPorts;
+    @Nullable private String workingDirectory;
 
     /**
      * Sets the image creation time.
@@ -70,7 +73,7 @@ public class Image<T extends Layer> {
      * @param value the value to set it to
      * @return this
      */
-    public Builder<T> setEnvironmentVariable(String name, String value) {
+    public Builder<T> addEnvironmentVariable(String name, String value) {
       environmentBuilder.put(name, value);
       return this;
     }
@@ -109,9 +112,9 @@ public class Image<T extends Layer> {
     }
 
     /**
-     * Add items to the "Labels" field in the container configuration.
+     * Adds items to the "Labels" field in the container configuration.
      *
-     * @param labels that map of labels to add
+     * @param labels the map of labels to add
      * @return this
      */
     public Builder<T> addLabels(@Nullable Map<String, String> labels) {
@@ -122,14 +125,25 @@ public class Image<T extends Layer> {
     }
 
     /**
-     * A an item to the "Labels" field in the container configuration.
+     * Adds an item to the "Labels" field in the container configuration.
      *
-     * @param name that name of the label
+     * @param name the name of the label
      * @param value the value of the label
      * @return this
      */
     public Builder<T> addLabel(String name, String value) {
       labelsBuilder.put(name, value);
+      return this;
+    }
+
+    /**
+     * Sets the item in the "WorkingDir" field in the container configuration.
+     *
+     * @param workingDirectory the working directory
+     * @return this
+     */
+    public Builder<T> setWorkingDirectory(@Nullable String workingDirectory) {
+      this.workingDirectory = workingDirectory;
       return this;
     }
 
@@ -145,15 +159,28 @@ public class Image<T extends Layer> {
       return this;
     }
 
+    /**
+     * Adds a history element to the image.
+     *
+     * @param history the history object to add
+     * @return this
+     */
+    public Builder<T> addHistory(HistoryEntry history) {
+      historyBuilder.add(history);
+      return this;
+    }
+
     public Image<T> build() {
       return new Image<>(
           created,
           imageLayersBuilder.build(),
+          historyBuilder.build(),
           environmentBuilder.build(),
           entrypoint,
           javaArguments,
           exposedPorts,
-          labelsBuilder.build());
+          labelsBuilder.build(),
+          workingDirectory);
     }
   }
 
@@ -166,6 +193,9 @@ public class Image<T extends Layer> {
 
   /** The layers of the image, in the order in which they are applied. */
   private final ImageLayers<T> layers;
+
+  /** The commands used to build each layer of the image */
+  private final ImmutableList<HistoryEntry> history;
 
   /** Environment variable definitions for running the image, in the format {@code NAME=VALUE}. */
   @Nullable private final ImmutableMap<String, String> environment;
@@ -182,21 +212,28 @@ public class Image<T extends Layer> {
   /** Labels on the container configuration */
   @Nullable private final ImmutableMap<String, String> labels;
 
+  /** Working directory on the container configuration */
+  @Nullable private final String workingDirectory;
+
   private Image(
       @Nullable Instant created,
       ImageLayers<T> layers,
+      ImmutableList<HistoryEntry> history,
       @Nullable ImmutableMap<String, String> environment,
       @Nullable ImmutableList<String> entrypoint,
       @Nullable ImmutableList<String> javaArguments,
       @Nullable ImmutableList<Port> exposedPorts,
-      @Nullable ImmutableMap<String, String> labels) {
+      @Nullable ImmutableMap<String, String> labels,
+      @Nullable String workingDirectory) {
     this.created = created;
     this.layers = layers;
+    this.history = history;
     this.environment = environment;
     this.entrypoint = entrypoint;
     this.javaArguments = javaArguments;
     this.exposedPorts = exposedPorts;
     this.labels = labels;
+    this.workingDirectory = workingDirectory;
   }
 
   @Nullable
@@ -229,7 +266,16 @@ public class Image<T extends Layer> {
     return labels;
   }
 
+  @Nullable
+  public String getWorkingDirectory() {
+    return workingDirectory;
+  }
+
   public ImmutableList<T> getLayers() {
     return layers.getLayers();
+  }
+
+  public ImmutableList<HistoryEntry> getHistory() {
+    return history;
   }
 }