+++ jib-core/src/main/java/com/google/cloud/tools/jib/image/json/ImageToJsonTranslator.java
@@ -115,6 +115,11 @@ public class ImageToJsonTranslator {
       template.addLayerDiffId(layer.getDiffId());
     }
 
+    // Adds the history.
+    for (HistoryEntry historyObject : image.getHistory()) {
+      template.addHistoryEntry(historyObject);
+    }
+
     // Sets the creation time. Instant#toString() returns an ISO-8601 formatted string.
     template.setCreated(image.getCreated() == null ? null : image.getCreated().toString());
 
@@ -132,6 +137,9 @@ public class ImageToJsonTranslator {
 
     // Sets the labels.
     template.setContainerLabels(image.getLabels());
+
+    // Sets the working directory.
+    template.setContainerWorkingDir(image.getWorkingDirectory());
 
     // Serializes into JSON.
     return JsonTemplateMapper.toBlob(template);