+++ jib-core/src/main/java/com/google/cloud/tools/jib/registry/credentials/DockerCredentialHelper.java
@@ -29,6 +29,7 @@ import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import javax.annotation.Nullable;
 
 /**
@@ -38,6 +39,8 @@ import javax.annotation.Nullable;
  *     href="https://github.com/docker/docker-credential-helpers">https://github.com/docker/docker-credential-helpers</a>
  */
 public class DockerCredentialHelper {
+
+  public static final String CREDENTIAL_HELPER_PREFIX = "docker-credential-";
 
   private final String serverUrl;
   private final Path credentialHelper;
@@ -51,14 +54,18 @@ public class DockerCredentialHelper {
   }
 
   /**
-   * Construct with {@link DockerCredentialHelperFactory}.
+   * Constructs a new {@link DockerCredentialHelper}.
    *
    * @param serverUrl the server URL to pass into the credential helper
    * @param credentialHelper the path to the credential helper executable
    */
-  DockerCredentialHelper(String serverUrl, Path credentialHelper) {
+  public DockerCredentialHelper(String serverUrl, Path credentialHelper) {
     this.serverUrl = serverUrl;
     this.credentialHelper = credentialHelper;
+  }
+
+  DockerCredentialHelper(String registry, String credentialHelperSuffix) {
+    this(registry, Paths.get(CREDENTIAL_HELPER_PREFIX + credentialHelperSuffix));
   }
 
   /**
@@ -69,13 +76,14 @@ public class DockerCredentialHelper {
    * }</pre>
    *
    * @return the Docker credentials by calling the corresponding CLI
-   * @throws IOException if writing/reading process input/output fails.
-   * @throws NonexistentServerUrlDockerCredentialHelperException if credentials are not found.
-   * @throws DockerCredentialHelperNotFoundException if the credential helper CLI doesn't exist.
+   * @throws IOException if writing/reading process input/output fails
+   * @throws CredentialHelperUnhandledServerUrlException if no credentials could be found for the
+   *     corresponding server
+   * @throws CredentialHelperNotFoundException if the credential helper CLI doesn't exist
    */
   public Credential retrieve()
-      throws IOException, NonexistentServerUrlDockerCredentialHelperException,
-          DockerCredentialHelperNotFoundException {
+      throws IOException, CredentialHelperUnhandledServerUrlException,
+          CredentialHelperNotFoundException {
     try {
       String[] credentialHelperCommand = {credentialHelper.toString(), "get"};
 
@@ -90,14 +98,14 @@ public class DockerCredentialHelper {
 
         // Throws an exception if the credential store does not have credentials for serverUrl.
         if (output.contains("credentials not found in native keychain")) {
-          throw new NonexistentServerUrlDockerCredentialHelperException(
+          throw new CredentialHelperUnhandledServerUrlException(
               credentialHelper, serverUrl, output);
         }
         if (output.isEmpty()) {
           try (InputStreamReader processStderrReader =
               new InputStreamReader(process.getErrorStream(), StandardCharsets.UTF_8)) {
             String errorOutput = CharStreams.toString(processStderrReader);
-            throw new NonexistentServerUrlDockerCredentialHelperException(
+            throw new CredentialHelperUnhandledServerUrlException(
                 credentialHelper, serverUrl, errorOutput);
           }
         }
@@ -107,14 +115,14 @@ public class DockerCredentialHelper {
               JsonTemplateMapper.readJson(output, DockerCredentialsTemplate.class);
           if (Strings.isNullOrEmpty(dockerCredentials.Username)
               || Strings.isNullOrEmpty(dockerCredentials.Secret)) {
-            throw new NonexistentServerUrlDockerCredentialHelperException(
+            throw new CredentialHelperUnhandledServerUrlException(
                 credentialHelper, serverUrl, output);
           }
 
           return new Credential(dockerCredentials.Username, dockerCredentials.Secret);
 
         } catch (JsonProcessingException ex) {
-          throw new NonexistentServerUrlDockerCredentialHelperException(
+          throw new CredentialHelperUnhandledServerUrlException(
               credentialHelper, serverUrl, output);
         }
       }
@@ -127,16 +135,11 @@ public class DockerCredentialHelper {
       // Checks if the failure is due to a nonexistent credential helper CLI.
       if (ex.getMessage().contains("No such file or directory")
           || ex.getMessage().contains("cannot find the file")) {
-        throw new DockerCredentialHelperNotFoundException(credentialHelper, ex);
+        throw new CredentialHelperNotFoundException(credentialHelper, ex);
       }
 
       throw ex;
     }
-  }
-
-  @VisibleForTesting
-  String getServerUrl() {
-    return serverUrl;
   }
 
   @VisibleForTesting