+++ jib-core/src/test/java/com/google/cloud/tools/jib/image/json/ImageToJsonTranslatorTest.java
@@ -56,8 +56,8 @@ public class ImageToJsonTranslatorTest {
     Image.Builder<CachedLayer> testImageBuilder = Image.builder();
 
     testImageBuilder.setCreated(Instant.ofEpochSecond(20));
-    testImageBuilder.setEnvironmentVariable("VAR1", "VAL1");
-    testImageBuilder.setEnvironmentVariable("VAR2", "VAL2");
+    testImageBuilder.addEnvironmentVariable("VAR1", "VAL1");
+    testImageBuilder.addEnvironmentVariable("VAR2", "VAL2");
     testImageBuilder.setEntrypoint(Arrays.asList("some", "entrypoint", "command"));
     testImageBuilder.setJavaArguments(Arrays.asList("arg1", "arg2"));
     testImageBuilder.setExposedPorts(
@@ -66,6 +66,7 @@ public class ImageToJsonTranslatorTest {
             new Port(2000, Protocol.TCP),
             new Port(3000, Protocol.UDP)));
     testImageBuilder.addLabels(ImmutableMap.of("key1", "value1", "key2", "value2"));
+    testImageBuilder.setWorkingDirectory("/some/workspace");
 
     DescriptorDigest fakeDigest =
         DescriptorDigest.fromDigest(
@@ -73,7 +74,19 @@ public class ImageToJsonTranslatorTest {
     CachedLayer fakeLayer =
         new CachedLayer(Mockito.mock(Path.class), new BlobDescriptor(1000, fakeDigest), fakeDigest);
     testImageBuilder.addLayer(fakeLayer);
-
+    testImageBuilder.addHistory(
+        HistoryEntry.builder()
+            .setCreationTimestamp(Instant.EPOCH)
+            .setAuthor("Bazel")
+            .setCreatedBy("bazel build ...")
+            .setEmptyLayer(true)
+            .build());
+    testImageBuilder.addHistory(
+        HistoryEntry.builder()
+            .setCreationTimestamp(Instant.ofEpochSecond(20))
+            .setAuthor("Jib")
+            .setCreatedBy("jib")
+            .build());
     imageToJsonTranslator = new ImageToJsonTranslator(testImageBuilder.build());
   }