+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/PluginConfigurationProcessor.java
@@ -88,7 +88,7 @@ class PluginConfigurationProcessor {
     if (fromCredential == null) {
       fromCredential = mavenSettingsServerCredentials.retrieve(baseImage.getRegistry());
       if (fromCredential != null) {
-        defaultCredentialRetrievers.setKnownCredential(
+        defaultCredentialRetrievers.setInferredCredential(
             fromCredential, MavenSettingsServerCredentials.CREDENTIAL_SOURCE);
       }
     } else {
@@ -120,6 +120,7 @@ class PluginConfigurationProcessor {
 
     BuildConfiguration.Builder buildConfigurationBuilder =
         BuildConfiguration.builder(logger)
+            .setCreatedBy("jib-maven-plugin")
             .setAllowInsecureRegistries(jibPluginConfiguration.getAllowInsecureRegistries())
             .setLayerConfigurations(
                 projectProperties.getJavaLayerConfigurations().getLayerConfigurations());